requirements.txt
diffusers==0.33.1
easydict==1.13
einops==0.8.1
h5py==3.14.0
lz4==4.4.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
msgpack==1.1.0
ninja==1.11.1.4
numpy==1.23.5
open3d==0.19.0
opencv-python==4.11.0.86
pandas==2.3.0
peft==0.15.2
pillow==11.0.0
scikit-learn==1.7.0
scipy==1.15.3
termcolor==3.1.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
huggingface_hub==0.33.0